producer/requirements.txt
pika
beautifulsoup4
lxml
requests
python-dotenv